separate requirement components
